refactor: move Tuples & Bearer token to utils (split package fix $52)
