+++ grpc/client/src/main/java/io/quarkiverse/authzed/utils/BearerToken.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.authzed;
+package io.quarkiverse.authzed.utils;
 
 import java.util.concurrent.Executor;
